+++ include/task_manager.h
@@ -57,9 +57,13 @@ namespace detail {
 		friend struct task_manager_testspy;
 
 	  public:
+		struct policy_set {
+			error_policy uninitialized_read_error = error_policy::throw_exception;
+		};
+
 		constexpr inline static task_id initial_epoch_task = 0;
 
-		task_manager(size_t num_collective_nodes, host_queue* queue, detail::task_recorder* recorder);
+		task_manager(size_t num_collective_nodes, host_queue* queue, detail::task_recorder* recorder, const policy_set& policy = default_policy_set());
 
 		virtual ~task_manager() = default;
 
@@ -189,6 +193,10 @@ namespace detail {
 		size_t get_current_task_count() const { return m_task_buffer.get_current_task_count(); }
 
 	  private:
+		// default-constructs a policy_set - this must be a function because we can't use the implicit default constructor of policy_set, which has member
+		// initializers, within its surrounding class (Clang)
+		constexpr static policy_set default_policy_set() { return {}; }
+
 		struct per_buffer_data {
 			std::string debug_name;
 
@@ -207,6 +215,7 @@ namespace detail {
 
 		const size_t m_num_collective_nodes;
 		host_queue* m_queue;
+		policy_set m_policy;
 
 		error_policy m_uninitialized_read_policy = error_policy::throw_exception;